% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_depth.R
\name{as_depth}
\alias{as_depth}
\title{Convert precipitation yield in precipitation depth as a function of duration.}
\usage{
as_depth(x = NULL, d = NULL)
}
\arguments{
\item{x}{numeric. Precipitation yield in l/(s*ha).}

\item{d}{numeric. Duration in minutes.}
}
\value{
units. Precipitation depth in mm.
}
\description{
Convert precipitation yield in precipitation depth as a function of duration.
}
\examples{
as_depth(126.94, d = 60)
}
