% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_height}
\alias{as_height}
\title{Convert precipitation yield in precipitation height as a function of duration.}
\usage{
as_height(rh = NULL, d = NULL)
}
\arguments{
\item{rh}{Precipitation yield in l/(s*ha).}

\item{d}{Duration in minutes.}
}
\value{
Precipitation height in mm.
}
\description{
Convert precipitation yield in precipitation height as a function of duration.
}
\examples{
as_yield(rh = 126.94, d = 60)
}
