% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_idx.R
\name{build_idx}
\alias{build_idx}
\title{Construct "INDEX_RC" based on given X and Y information}
\usage{
build_idx(col, row)
}
\arguments{
\item{col}{Integer {1:79}}

\item{row}{Integer {1:107}}
}
\value{
A string containing the the unique representation of the relevant
"INDEX_RC" field.
}
\description{
Construct "INDEX_RC" based on given X and Y information
}
\examples{
\dontrun{
build_idx(11, 49)
}
}
