% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_table.R
\name{build_table}
\alias{build_table}
\title{Get cell-specific statistics from the KOSTRA-2010R data set}
\usage{
build_table(grid_index)
}
\arguments{
\item{grid_index}{A String representing the relevant "INDEX_RC" field.}
}
\value{
A tibble containing precipitation heights as a function of duration
and return period for the KOSTRA-2010R grid cell specified.
}
\description{
Get cell-specific statistics from the KOSTRA-2010R data set
}
\examples{
\dontrun{
build_table("49011")
}
}
