% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_designstorm.R
\name{calc_designstorm}
\alias{calc_designstorm}
\title{Design storm calculation based on statistical precipitation depths}
\usage{
calc_designstorm(data = NULL, tn = NULL, d = NULL, type = NULL)
}
\arguments{
\item{data}{A tibble containing grid cell statistics from KOSTRA-DWD-2010R.}

\item{tn}{Return periods in years.}

\item{d}{Duration in minutes.}

\item{type}{Precipitation distribution: EulerI | EulerII.}
}
\value{
An xts object.
}
\description{
Design storm calculation based on statistical precipitation depths
}
\examples{
kostra <- get_stats("49011")
xts <- calc_designstorm(kostra, tn = 20, d = 60, type = "EulerII")
}
