% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_modelrain.R
\name{calc_modelrain}
\alias{calc_modelrain}
\title{Calculate model rainfall from statistical precipitation}
\usage{
calc_modelrain(data, d, tn, type = "EulerII")
}
\arguments{
\item{data}{A tibble containing grid cell statistics from KOSTRA-2010R.}

\item{d}{Duration in minutes.}

\item{tn}{Return periods in years.}

\item{type}{EulerI | EulerII.}
}
\value{
An xts object.
}
\description{
Calculate model rainfall from statistical precipitation
}
\examples{
\dontrun{
xts <- calc_modelrain(kostra, d = 60, tn = 20, type = "EulerII")
}
}
