% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pen.R
\name{calc_pen}
\alias{calc_pen}
\title{Extrapolation of precipitation heights for Tn > 100 a according to PEN-LAWA}
\usage{
calc_pen(tibble)
}
\arguments{
\item{tibble}{Cell-based statistics derived from KOSTRA-2010R}
}
\value{
A tibble containing extrapolated precipitation heights as a function
of duration and return periods.
}
\description{
Extrapolation of precipitation heights for Tn > 100 a according to PEN-LAWA
}
\examples{
\dontrun{
calc_pen(kostra)
}
}
\references{
Verworn & Draschoff (2006)

Verworn & Kummer (2008)
}
