% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_centroid}
\alias{get_centroid}
\title{Construct an sf object of type point using coordinates or specific polygons}
\usage{
get_centroid(input, crs = 25832)
}
\arguments{
\item{input}{Vector of length 2 containing numeric representing coordinates,
string of length 1 representing the name of a municipality,
or string of length 5 representing a postal zip code.}

\item{crs}{(optional) Coordinate reference system definition.}
}
\value{
An object of type \code{sfc_POINT}.
}
\description{
Construct an sf object of type point using coordinates or specific polygons
}
\examples{
\dontrun{
p1 <- get_centroid(input = c(367773, 5703579))
p2 <- get_centroid(input = c(6.09, 50.46), crs = 4326)
p3 <- get_centroid(input = "Essen")
p4 <- get_centroid(input = "45145")
}
}
