% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_depth.R
\name{get_depth}
\alias{get_depth}
\title{Get precipitation depth for a specified duration and return period}
\usage{
get_depth(x = NULL, d = NULL, tn = NULL)
}
\arguments{
\item{x}{Tibble containing grid cell statistics from KOSTRA-2010R.}

\item{d}{numeric. Precipitation duration level \code{[min]}.}

\item{tn}{numeric. Return period \code{[a]}.}
}
\value{
units. Precipitation depth \code{[mm]}.
}
\description{
Get precipitation depth for a specified duration and return period
}
\examples{
kostra <- get_stats("49011")

get_depth(kostra, d = 60, tn = 50)
}
