% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_idx.R
\name{get_idx}
\alias{get_idx}
\title{Get index of KOSTRA-2010R grid by means of intersection with given location}
\usage{
get_idx(location)
}
\arguments{
\item{location}{Sf object containing a point feature.}
}
\value{
A string containing the the unique representation of the relevant
"INDEX_RC" field.
}
\description{
Get index of KOSTRA-2010R grid by means of intersection with given location
}
\examples{
\dontrun{
get_idx(p)
}
}
