% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Get cell-specific parameters from KOSTRA-DWD-2020 dataset}
\usage{
get_params(x = NULL)
}
\arguments{
\item{x}{character. Relevant "INDEX_RC" field to be queried.}
}
\value{
Tibble containing estimated GEV parameters for the KOSTRA-DWD-2020
grid cell specified.
}
\description{
Get cell-specific parameters from KOSTRA-DWD-2020 dataset
}
\examples{
get_params("117111")
}
\seealso{
\code{\link[=idx_build]{idx_build()}}
}
