% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_returnp.R
\name{get_returnp}
\alias{get_returnp}
\title{Get return period class for specified precipitation depth}
\usage{
get_returnp(data = NULL, hn = NULL, d = NULL)
}
\arguments{
\item{data}{A tibble containing grid cell statistics from KOSTRA-2010R.}

\item{hn}{Precipitation depth in mm.}

\item{d}{Duration in minutes.}
}
\value{
A numerical vector representing the upper and lower boundaries of the
return period class in years.
}
\description{
Get return period class for specified precipitation depth
}
\examples{
get_returnp(kostra, hn = 69.3, d = 1440)
}
