% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{Get cell-specific statistics from the KOSTRA-2010R data set}
\usage{
get_stats(x = NULL)
}
\arguments{
\item{x}{character. Relevant "INDEX_RC" field to be queried.}
}
\value{
Tibble containing statistical precipitation depths as a function of
duration and return periods for the KOSTRA-2010R grid cell specified.
}
\description{
Get cell-specific statistics from the KOSTRA-2010R data set
}
\examples{
kostra <- get_stats("49011")
}
