% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_idf.R
\name{ggplot_idf}
\alias{ggplot_idf}
\title{Visualize intensity-duration-frequency curves per cell-specific statistics}
\usage{
ggplot_idf(x = NULL, tn = NULL, log10 = FALSE)
}
\arguments{
\item{x}{Tibble containing grid cell statistics from KOSTRA-DWD-2020,
as provided by \code{get_stats()}.}

\item{tn}{(optional) numeric. Return period \code{[a]} to be used for filtering.}

\item{log10}{logical. Transform x axis to log10 scale?}
}
\value{
ggplot object.
}
\description{
Visualize intensity-duration-frequency curves per cell-specific statistics
}
\examples{
\dontrun{
get_stats("117111") |> ggplot_idf()

get_stats("117111") |> ggplot_idf(tn = 100)

get_stats("117111") |> ggplot_idf(log10 = TRUE)

get_stats("117111", as_depth = FALSE) |> ggplot_idf()
}
}
\seealso{
\code{\link[=get_stats]{get_stats()}}
}
