% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_spatial.R
\name{ggplot_spatial}
\alias{ggplot_spatial}
\title{Visualize the spatial distribution of cell-specific statistical precipitation depths}
\usage{
ggplot_spatial(d = NULL, tn = NULL)
}
\arguments{
\item{d}{numeric. Precipitation duration level \code{[min]}.}

\item{tn}{numeric. Return period \code{[a]}.}
}
\value{
ggplot object.
}
\description{
Visualize the spatial distribution of cell-specific statistical precipitation depths
}
\examples{
\dontrun{
ggplot_spatial(d = 60, tn = 1)

ggplot_spatial(d = 1440, tn = 100)
}
}
\seealso{
\code{\link[=get_stats]{get_stats()}}
}
