% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_decompose.R
\name{idx_decompose}
\alias{idx_decompose}
\title{Extract row and column information from "INDEX_RC" identifier}
\usage{
idx_decompose(x = NULL)
}
\arguments{
\item{x}{character. Unique representation of the relevant "INDEX_RC" field.}
}
\value{
integer. Vector of length 2 containing row and column numbers.
}
\description{
Extract row and column information from "INDEX_RC" identifier
}
\examples{
idx_decompose(x = "49011")
}
\seealso{
\code{\link[=idx_build]{idx_build()}}
}
