% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_exists.R
\name{idx_exists}
\alias{idx_exists}
\title{Check whether "INDEX_RC" provided is present in KOSTRA-2010R data set}
\usage{
idx_exists(idx = NULL)
}
\arguments{
\item{idx}{character. Relevant "INDEX_RC" field to be queried.}
}
\value{
logical.
}
\description{
Check whether "INDEX_RC" provided is present in KOSTRA-2010R data set
}
\examples{
idx_exists("49011")
}
