% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_utils.R
\name{idx_get}
\alias{idx_get}
\title{Get index of KOSTRA-2010R cell by means of intersection with given location}
\usage{
idx_get(location)
}
\arguments{
\item{location}{Sf object containing a point feature.}
}
\value{
A string containing the unique representation of the relevant
"INDEX_RC" field.
}
\description{
Get index of KOSTRA-2010R cell by means of intersection with given location
}
\examples{
\dontrun{
idx_get(p)
}
}
