% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kostra_dwd_2010r}
\alias{kostra_dwd_2010r}
\title{KOSTRA-DWD-2010R}
\format{
A list of length 18 with simple feature collections of type POLYGON with 8,453 features and 10 field each:
\describe{
\item{INDEX_RC}{grid index}
\item{HN_001A_}{design precipitation for return period 1 a}
\item{HN_002A_}{design precipitation for return period 2 a}
\item{HN_003A_}{design precipitation for return period 3 a}
\item{HN_005A_}{design precipitation for return period 5 a}
\item{HN_010A_}{design precipitation for return period 10 a}
\item{HN_020A_}{design precipitation for return period 20 a}
\item{HN_030A_}{design precipitation for return period 30 a}
\item{HN_050A_}{design precipitation for return period 50 a}
\item{HN_100A_}{design precipitation for return period 100 a}
\item{geometry}{polygon coordinates}
}
}
\source{
\url{https://opendata.dwd.de/climate_environment/CDC/grids_germany/return_periods/precipitation/KOSTRA/KOSTRA_DWD_2010R/gis/}
}
\usage{
kostra_dwd_2010r
}
\description{
\url{https://www.dwd.de/DE/leistungen/kostra_dwd_rasterwerte/kostra_dwd_rasterwerte.html}
}
\details{
Statistical precipitation values as a function of duration and return period
}
\note{
Last access: 2022-10-20

License: Data licence Germany – attribution – version 2.0

Copyright: Deutscher Wetterdienst 2022 (format modified)
}
\keyword{datasets}
