% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kostra_dwd_2020_params}
\alias{kostra_dwd_2020_params}
\title{Data: KOSTRA-DWD-2020 parameters}
\format{
A tibble with 15,989 rows and 6 columns:
\describe{
\item{INDEX_RC}{numeric. Unique grid cell identifier.}

\item{XI}{numeric. GEV location parameter.}
\item{ALPHA}{numeric. GEV scale parameter.}
\item{KAPPA}{numeric. GEV shape parameter.}
\item{THETA}{numeric. First Koutsoyiannis parameter.}
\item{ETA}{numeric. Second Koutsoyiannis parameter.}
}
}
\source{
\url{https://opendata.dwd.de/climate_environment/CDC/grids_germany/return_periods/precipitation/KOSTRA/KOSTRA_DWD_2020/tab/}
}
\usage{
kostra_dwd_2020_params
}
\description{
\url{https://www.dwd.de/DE/leistungen/kostra_dwd_rasterwerte/kostra_dwd_rasterwerte.html}
}
\details{
Estimated GEV parameters per KOSTRA-DWD-2020 grid cell to reconstruct underlying extreme value distributions
}
\note{
Last access: 2025-02-25

License: CC BY 4.0

Copyright: Deutscher Wetterdienst 2022 (format modified)
}
\keyword{datasets}
