% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{osm_plz}
\alias{osm_plz}
\title{Centroids of zip code areas in Germany}
\format{
Simple feature collection of type POINT with 8,170 features and 1 field:
\describe{
\item{plz}{zip codes}
\item{geometry}{centroid coordinates}
}
}
\source{
\url{https://www.suche-postleitzahl.org/downloads}
}
\usage{
osm_plz
}
\description{
A subset/derivate of 5-digit zip code data provided by OpenStreetMap
}
\details{
: osm_plz <- sf::st_read("plz-5stellig.shp") |> dplyr::select("plz") |> sf::st_centroid() |> sf::st_transform("epsg:25832")
}
\note{
: License: Open Data Commons Open Database License (ODbL)

: Copyright: OpenStreetMap contributors 2022 (modified)
}
\keyword{datasets}
