% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{osm_plz_centroids}
\alias{osm_plz_centroids}
\title{Data: Centroids of zip code areas in Germany}
\format{
Simple feature collection of type POINT with 8,170 features and 1 field:
\describe{
\item{plz}{zip codes}
\item{geometry}{centroid coordinates}
}
}
\source{
\url{https://downloads.suche-postleitzahl.org/v2/public/plz-5stellig.shp.zip}
}
\usage{
osm_plz_centroids
}
\description{
\url{https://www.suche-postleitzahl.org/downloads}
}
\details{
A subset/derivative of 5-digit zip code data provided by OpenStreetMap
}
\note{
Last access: 2022-10-20

License: Open Data Commons Open Database License (ODbL)

Copyright: OpenStreetMap contributors 2022 (modified)
}
\keyword{datasets}
