% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vg250_gem_centroids}
\alias{vg250_gem_centroids}
\title{Centroids of administrative areas on municipality level in Germany (VG250 31.12.)}
\format{
Simple feature collection with 11,123 features and 1 field:
\describe{
\item{GEN}{municipality name}
\item{geometry}{centroid coordinates}
}
}
\source{
\url{https://gdz.bkg.bund.de/index.php/default/digitale-geodaten/verwaltungsgebiete/verwaltungsgebiete-1-250-000-stand-31-12-vg250-31-12.html}
}
\usage{
vg250_gem_centroids
}
\description{
A subset/derivate of data from the VG250_GEM product provided by the Federal Agency for Cartography and Geodesy, Germany
}
\details{
: sf::st_read("VG250_GEM.shp") |> dplyr::select("GEN") |> sf::st_centroid() |> sf::st_transform("epsg:25832")
}
\note{
: License: Datenlizenz Deutschland - Namensnennung - Version 2.0

: Copyright: GeoBasis-DE / BKG 2022 (modified)
}
\keyword{datasets}
