% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vg250_pk}
\alias{vg250_pk}
\title{Point representation of municipalities in Germany (VG250 31.12.)}
\format{
Simple feature collection with 10,994 features and 1 field:
\describe{
\item{GEN}{municipality name}
\item{geometry}{centroid coordinates}
}
}
\source{
\url{https://gdz.bkg.bund.de/index.php/default/digitale-geodaten/verwaltungsgebiete/verwaltungsgebiete-1-250-000-stand-31-12-vg250-31-12.html}
}
\usage{
vg250_pk
}
\description{
A subset of data from the VG250_PK product provided by the Federal Agency for Cartography and Geodesy, Germany
}
\details{
: sf::st_read("VG250_PK.shp") |> dplyr::select("GEN") |> sf::st_transform("epsg:25832")
}
\note{
: License: Data licence Germany – attribution – version 2.0

: Copyright: GeoBasis-DE / BKG 2022 (modified)
}
\keyword{datasets}
