% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stats.R
\name{write_stats}
\alias{write_stats}
\title{Write cell-specific statistics from KOSTRA-DWD-2020 dataset to disk}
\usage{
write_stats(x = NULL, file = NULL)
}
\arguments{
\item{x}{Tibble containing grid cell statistics from KOSTRA-DWD-2020,
as provided by \code{get_stats()}.}

\item{file}{(optional) character. Filename with csv extension to be used for data output.}
}
\description{
Write cell-specific statistics from KOSTRA-DWD-2020 dataset to disk
}
\examples{
\dontrun{
stats <- get_stats("117111")

write_stats(stats)
write_stats(stats, file = "kostra2020_hN_117111.csv")
}
}
\seealso{
\code{\link[=get_stats]{get_stats()}}
}
