% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_depth.R
\name{as_depth}
\alias{as_depth}
\title{Convert precipitation yield in precipitation depth as a function of duration}
\usage{
as_depth(x = NULL, d = NULL)
}
\arguments{
\item{x}{numeric. Precipitation yield \code{[l s-1 ha-1]}.}

\item{d}{numeric. Precipitation duration level \code{[min]}.}
}
\value{
units. Precipitation depth \code{[mm]}.
}
\description{
Convert precipitation yield in precipitation depth as a function of duration
}
\examples{
as_depth(x = 126.94, d = 60)
}
\seealso{
\link{as_yield}
}
