% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_yield}
\alias{as_yield}
\title{Convert precipitation depth in precipitation yield as a function of duration.}
\usage{
as_yield(hn = NULL, d = NULL)
}
\arguments{
\item{hn}{Precipitation depth in mm.}

\item{d}{Duration in minutes.}
}
\value{
Precipitation yield in l/(s*ha).
}
\description{
Convert precipitation depth in precipitation yield as a function of duration.
}
\examples{
as_yield(hn = 45.7, d = 60)
}
