% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_yield.R
\name{as_yield}
\alias{as_yield}
\title{Convert precipitation depth in precipitation yield as a function of duration.}
\usage{
as_yield(x = NULL, d = NULL)
}
\arguments{
\item{x}{numeric. Precipitation depth in mm.}

\item{d}{numeric. Duration in minutes.}
}
\value{
units. Precipitation yield in l/(s*ha).
}
\description{
Convert precipitation depth in precipitation yield as a function of duration.
}
\examples{
as_yield(45.7, d = 60)
}
