% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_yield.R
\name{as_yield}
\alias{as_yield}
\title{Convert precipitation depth in precipitation yield as a function of duration}
\usage{
as_yield(x = NULL, d = NULL)
}
\arguments{
\item{x}{numeric. Precipitation depth \code{[mm]}.}

\item{d}{numeric. Precipitation duration level \code{[min]}.}
}
\value{
units. Precipitation yield \code{[l s-1 ha-1]}.
}
\description{
Convert precipitation depth in precipitation yield as a function of duration
}
\examples{
as_yield(x = 45.7, d = 60)
}
\seealso{
\link{as_depth}
}
