% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_table.R
\name{build_table}
\alias{build_table}
\title{Get interval/frequency statistics for the KOSTRA-2010R grid specified}
\usage{
build_table(grid_index)
}
\arguments{
\item{grid_index}{A String representing the relevant "INDEX_RC" field.}
}
\value{
A tibble containing precipitation height interval/frequency
statistics for a defined KOSTRA-2010R grid.
}
\description{
Get interval/frequency statistics for the KOSTRA-2010R grid specified
}
\examples{
\dontrun{
build_table("49011")
}
}
