% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_designstorm.R
\name{calc_designstorm}
\alias{calc_designstorm}
\title{Design storm calculation based on statistical precipitation}
\usage{
calc_designstorm(data, d, tn, type = "EulerII")
}
\arguments{
\item{data}{A tibble containing grid cell statistics from KOSTRA-DWD-2010R.}

\item{d}{Duration in minutes.}

\item{tn}{Return periods in years.}

\item{type}{EulerI | EulerII.}
}
\value{
An xts object.
}
\description{
Design storm calculation based on statistical precipitation
}
\examples{
\dontrun{
xts <- calc_designstorm(kostra, d = 60, tn = 20, type = "EulerII")
}
}
