% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_designstorm.R
\name{calc_designstorm}
\alias{calc_designstorm}
\title{Design storm calculation based on statistical precipitation depths}
\usage{
calc_designstorm(x = NULL, d = NULL, tn = NULL, type = NULL)
}
\arguments{
\item{x}{Tibble containing grid cell statistics from KOSTRA-DWD-2020,
as provided by \code{get_stats()}.}

\item{d}{numeric. Precipitation duration level \code{[min]}.}

\item{tn}{numeric. Return period \code{[a]}.}

\item{type}{character. Precipitation distribution: "EulerI" or "EulerII".}
}
\value{
An xts object.
}
\description{
Design storm calculation based on statistical precipitation depths
}
\examples{
kostra <- get_stats("49125")

calc_designstorm(kostra, d = 60, tn = 20, type = "EulerII")
}
\seealso{
\code{\link[=get_stats]{get_stats()}}
}
