% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_modelrain.R
\name{calc_modelrain}
\alias{calc_modelrain}
\title{Calculate model rainfall from statistical precipitation}
\usage{
calc_modelrain(tibble, d, tn, type = "EulerII")
}
\arguments{
\item{tibble}{containing statistics}

\item{d}{duration in minutes}

\item{tn}{return periods in years}

\item{type}{EulerI | EulerII}
}
\value{
xts
}
\description{
Calculate model rainfall from statistical precipitation
}
\examples{
\dontrun{
xts <- calc_modelrain(kostra, d = 60, tn = 100, type = "EulerII")
}
}
\references{
EulerI + EulerII
}
