% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pen.R
\name{calc_pen}
\alias{calc_pen}
\title{Extrapolation of precipitation depths for Tn > 100 a according to PEN-LAWA}
\usage{
calc_pen(x = NULL)
}
\arguments{
\item{x}{Tibble containing grid cell statistics from KOSTRA-2010R.}
}
\value{
Tibble containing extrapolated precipitation depths as a function of
duration and return periods.
}
\description{
Extrapolation of precipitation depths for Tn > 100 a according to PEN-LAWA
}
\examples{
kostra <- get_stats("49011")
pen <- calc_pen(kostra)

}
\references{
Verworn & Kummer (2006): Praxisrelevante Extremwerte des Niederschlages (PEN).

Verworn & Draschoff (2008): PEN-Erweiterung.
}
