% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pen.R
\name{calc_pen}
\alias{calc_pen}
\title{Extrapolation of precipitation heights for Tn > 100 a according to PEN-LAWA}
\usage{
calc_pen(data = NULL)
}
\arguments{
\item{data}{A tibble containing grid cell statistics from KOSTRA-2010R.}
}
\value{
A tibble containing extrapolated precipitation heights as a function
of duration and return periods.
}
\description{
Extrapolation of precipitation heights for Tn > 100 a according to PEN-LAWA
}
\examples{
\dontrun{
pen <- calc_pen(kostra)
}
}
\references{
Verworn & Kummer (2006): Praxisrelevante Extremwerte des Niederschlages (PEN).

Verworn & Draschoff (2008): PEN-Erweiterung.
}
