% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_centroid.R
\name{get_centroid}
\alias{get_centroid}
\title{Construct a geometry set object of type sfc_POINT based on user input}
\usage{
get_centroid(x = NULL, epsg = 25832)
}
\arguments{
\item{x}{Vector of length 2 containing numeric representing coordinates, \cr
or character of length 1 representing the name of a municipality, \cr
or character of nchar 5 representing a postal zip code.}

\item{epsg}{numeric. Coordinate reference system identifier.}
}
\value{
Object of type \code{sfc_POINT}.
}
\description{
Construct a geometry set object of type sfc_POINT based on user input
}
\examples{
get_centroid(c(367773, 5703579))
get_centroid(c(6.09, 50.46), epsg = 4326)
get_centroid("Aachen")
get_centroid("52070")
}
