% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_idx}
\alias{get_idx}
\title{Get index of relevant KOSTRA-2010R cell based on user-defined geometry}
\usage{
get_idx(location = NULL)
}
\arguments{
\item{location}{Object of type \code{sfc_POINT}, as provided by \code{get_centroid()}.}
}
\value{
character. Unique representation of the relevant "INDEX_RC" field.
}
\description{
Get index of relevant KOSTRA-2010R cell based on user-defined geometry
}
\examples{
p <- get_centroid(x = c(367773, 5703579))
get_idx(p)
}
