% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_idx}
\alias{get_idx}
\title{Get index of KOSTRA-2010R cell by means of intersection with given location}
\usage{
get_idx(location = NULL)
}
\arguments{
\item{location}{An object of type \code{sfc_POINT}.}
}
\value{
A string containing the unique representation of the relevant
"INDEX_RC" field.
}
\description{
Get index of KOSTRA-2010R cell by means of intersection with given location
}
\examples{
p <- get_centroid(input = c(367773, 5703579))
get_idx(p)
}
