% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_idx.R
\name{get_idx}
\alias{get_idx}
\title{Get relevant "INDEX_RC" based on user-defined geometry}
\usage{
get_idx(x = NULL)
}
\arguments{
\item{x}{Object of type \code{sfc_POINT}, as provided by \code{get_centroid()}.}
}
\value{
character. Unique representation of the relevant "INDEX_RC" field.
}
\description{
Get relevant "INDEX_RC" based on user-defined geometry
}
\examples{
p <- get_centroid(c(367773, 5703579))

get_idx(p)
}
\seealso{
\link{get_centroid}
}
