% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pdepth.R
\name{get_pdepth}
\alias{get_pdepth}
\title{Get precipitation depth for a specified duration and return period}
\usage{
get_pdepth(data = NULL, d = NULL, tn = NULL)
}
\arguments{
\item{data}{Tibble containing grid cell statistics from KOSTRA-2010R.}

\item{d}{numeric. Duration in minutes.}

\item{tn}{numeric. Return period in years.}
}
\value{
units. Precipitation depth in mm.
}
\description{
Get precipitation depth for a specified duration and return period
}
\examples{
kostra <- get_stats("49011")
get_pdepth(kostra, d = 60, tn = 50)
}
