% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_returnp.R
\name{get_returnp}
\alias{get_returnp}
\title{Get return period class for specified precipitation depth}
\usage{
get_returnp(data = NULL, hn = NULL, d = NULL)
}
\arguments{
\item{data}{Tibble containing grid cell statistics from KOSTRA-2010R.}

\item{hn}{numeric. Precipitation depth in mm.}

\item{d}{numeric. Duration in minutes.}
}
\value{
units. Vector of length 2 representing the upper and lower boundaries
of the return period class in years.
}
\description{
Get return period class for specified precipitation depth
}
\examples{
kostra <- get_stats("49011")
get_returnp(kostra, hn = 69.3, d = 1440)
}
