% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_returnp.R
\name{get_returnp}
\alias{get_returnp}
\title{Get return period for specified duration level and precipitation depth}
\usage{
get_returnp(x = NULL, hn = NULL, d = NULL, interpolate = FALSE)
}
\arguments{
\item{x}{Tibble containing grid cell statistics from KOSTRA-DWD-2020,
as provided by \code{get_stats()}.}

\item{hn}{numeric. Precipitation depth \code{[mm]}.}

\item{d}{numeric. Precipitation duration level \code{[min]}.}

\item{interpolate}{logical.
Return \code{tn} as an interpolated value instead of an interval?}
}
\value{
units. Vector of length 2 representing the upper and lower boundaries
of the return period class in years \code{[a]}. \cr
Vector of length 1 with \code{interpolate = TRUE}.
}
\description{
Get return period for specified duration level and precipitation depth
}
\examples{
stats <- get_stats("117111")

get_returnp(stats, hn = 69.3, d = 1440)
get_returnp(stats, hn = 69.3, d = 1440, interpolate = TRUE)
}
\seealso{
\code{\link[=get_stats]{get_stats()}}
}
