% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{Get cell-specific statistics from the KOSTRA-DWD-2020 dataset}
\usage{
get_stats(x = NULL, as_depth = TRUE)
}
\arguments{
\item{x}{character. Relevant "INDEX_RC" field to be queried.}

\item{as_depth}{logical. Returns precipitation depths when \code{TRUE} and precipitation
yields when \code{FALSE}.}
}
\value{
Tibble containing statistical precipitation depths as a function of
duration and return periods for the KOSTRA-DWD-2020 grid cell specified.
}
\description{
Get cell-specific statistics from the KOSTRA-DWD-2020 dataset
}
\examples{
get_stats("49125")
get_stats("49125", as_depth = FALSE)
}
