% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{Get cell-specific statistics from the KOSTRA-2010R data set}
\usage{
get_stats(grid_index)
}
\arguments{
\item{grid_index}{A String representing the relevant "INDEX_RC" field.}
}
\value{
A tibble containing precipitation heights as a function of duration
and return periods for the KOSTRA-2010R grid cell specified.
}
\description{
Get cell-specific statistics from the KOSTRA-2010R data set
}
\examples{
\dontrun{
kdata <- get_stats("49011")
}
}
