% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{idx_build}
\alias{idx_build}
\title{Construct "INDEX_RC" based on given X and Y information}
\usage{
idx_build(col = NULL, row = NULL)
}
\arguments{
\item{col}{integer. One pick from set {0:78}}

\item{row}{integer. One pick from set {0:106}}
}
\value{
character. Unique representation of the relevant "INDEX_RC" field.
}
\description{
Construct "INDEX_RC" based on given X and Y information
}
\examples{
idx_build(col = 11, row = 49)
}
