% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_build.R
\name{idx_build}
\alias{idx_build}
\title{Build "INDEX_RC" identifier based on given row and column information}
\usage{
idx_build(row = NULL, col = NULL)
}
\arguments{
\item{row}{integer. Row number, one pick from set {0:299}.}

\item{col}{integer. Column number, one pick from set {0:299}.}
}
\value{
character. Unique representation of the relevant "INDEX_RC" field.
}
\description{
Build "INDEX_RC" identifier based on given row and column information
}
\examples{
idx_build(row = 117, col = 111)
}
\seealso{
\code{\link[=idx_decompose]{idx_decompose()}}
}
