% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_exists.R
\name{idx_exists}
\alias{idx_exists}
\title{Check whether defined index is present in KOSTRA-2010R}
\usage{
idx_exists(idx)
}
\arguments{
\item{idx}{A string representing the grid cell index.}
}
\value{
A boolean.
}
\description{
Check whether defined index is present in KOSTRA-2010R
}
\examples{
\dontrun{
idx_exists("49011")
}
}
