% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_exists.R
\name{idx_exists}
\alias{idx_exists}
\title{Check whether given "INDEX_RC" is present in KOSTRA-DWD-2020 dataset}
\usage{
idx_exists(x = NULL)
}
\arguments{
\item{x}{character. Relevant "INDEX_RC" field to be queried.}
}
\value{
logical.
}
\description{
Check whether given "INDEX_RC" is present in KOSTRA-DWD-2020 dataset
}
\examples{
idx_exists("117111")
}
