% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{idx_exists}
\alias{idx_exists}
\title{Check whether "INDEX_RC" provided is present in KOSTRA-2010R data set}
\usage{
idx_exists(idx = NULL)
}
\arguments{
\item{idx}{A string representing the grid cell index.}
}
\value{
A boolean.
}
\description{
Check whether "INDEX_RC" provided is present in KOSTRA-2010R data set
}
\examples{
\dontrun{
idx_exists("49011")
}
}
