% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_idf.R
\name{plot_idf}
\alias{plot_idf}
\title{Visualize intensity-duration-frequency curves per cell-specific statistics from KOSTRA-DWD-2020 dataset}
\usage{
plot_idf(x = NULL, tn = NULL, log10 = FALSE)
}
\arguments{
\item{x}{Tibble containing grid cell statistics from KOSTRA-DWD-2020,
as provided by \code{get_stats()}.}

\item{tn}{(optional) numeric. Return period \code{[a]} to be used for filtering.}

\item{log10}{logical. Transform x axis to log10 scale?}
}
\value{
ggplot object.
}
\description{
Visualize intensity-duration-frequency curves per cell-specific statistics from KOSTRA-DWD-2020 dataset
}
\examples{
get_stats("49125") |> plot_idf()

get_stats("49125") |> plot_idf(tn = 100)

get_stats("49125") |> plot_idf(log10 = TRUE)
}
\seealso{
\code{\link[=get_stats]{get_stats()}}
}
