% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_idf.R
\name{plot_idf}
\alias{plot_idf}
\title{Visualize intensity-duration-frequency curves per grid cell statistics}
\usage{
plot_idf(x = NULL)
}
\arguments{
\item{x}{Tibble containing grid cell statistics from KOSTRA-DWD-2010R,
as provided by \code{get_stats()}.}
}
\value{
ggplot object.
}
\description{
Visualize intensity-duration-frequency curves per grid cell statistics
}
\examples{
get_stats("49125") |> plot_idf()
}
\seealso{
\code{\link[=get_stats]{get_stats()}}
}
