% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vg250_pk}
\alias{vg250_pk}
\title{Data: Point representation of municipalities in Germany (VG250 31.12.2021)}
\format{
Simple feature collection of type POINT with 10,994 features and 1 field:
\describe{
\item{GEN}{municipality name}
\item{geometry}{centroid coordinates}
}
}
\source{
\url{https://daten.gdz.bkg.bund.de/produkte/vg/vg250_ebenen_1231/aktuell/vg250_12-31.utm32s.shape.ebenen.zip}
}
\usage{
vg250_pk
}
\description{
\url{https://gdz.bkg.bund.de/index.php/default/digitale-geodaten/verwaltungsgebiete/verwaltungsgebiete-1-250-000-stand-31-12-vg250-31-12.html}
}
\details{
A subset of data from the VG250_PK product provided by the Federal Agency for Cartography and Geodesy, Germany
}
\note{
Last access: 2022-10-20

License: Data licence Germany – attribution – version 2.0

Copyright: GeoBasis-DE / BKG 2022 (modified)
}
\keyword{datasets}
